requirements.txt
requests
python-telegram-bot
loguru
pyyaml
configparser
